+++ public/sw.js
@@ -1,4 +1,4 @@
-const CACHE = 'vbm-cinematic-v4'
+const CACHE = 'vbm-manager-depth-v5'
 self.addEventListener('install', (event) => {
   event.waitUntil(caches.open(CACHE).then((cache) => cache.addAll(['./', './index.html', './manifest.webmanifest'])))
   self.skipWaiting()